64-bit Base64, HexDump
These routines encode/decode a block of binary data to/from a readable string.
The size of such a block was limited to 32-bit, now it's 64-bit.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13081 7512d703-071c-0410-b5c2-d2bdd895b7d1

+++ src/utils/Base64.hh
@@ -6,7 +6,7 @@
 #include <string>
 
 namespace Base64 {
-	std::string encode(const void* input, int len);
+	std::string encode(const void* input, size_t len);
 	std::string decode(const std::string& input);
 }
